+++ airdnb/src/main/java/team10/airdnb/config/global/SecurityConfig.java
@@ -1,9 +1,9 @@
-package team10.airdnb.oauth.config;
+package team10.airdnb.config.global;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import team10.airdnb.oauth.jwt.service.TokenManager;
+import team10.airdnb.jwt.service.TokenManager;
 
 //jwt yml 설정 파일 읽기
 @Configuration
