+++ airdnb/src/main/java/team10/airdnb/exception/GlobalExceptionHandler.java
@@ -6,19 +6,31 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import team10.airdnb.exception.response.BusinessExceptionResponse;
 import team10.airdnb.exception.response.ValidExceptionResponse;
+import team10.airdnb.member.exception.MemberIdNotFoundException;
+import team10.airdnb.oauth.exception.AuthenticationException;
 
 @Slf4j
 @RestControllerAdvice
 public class GlobalExceptionHandler {
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<?> handleValidationException(MethodArgumentNotValidException e) {
+    public ResponseEntity<ValidExceptionResponse> handleValidationException(MethodArgumentNotValidException e) {
         ValidExceptionResponse response = ValidExceptionResponse.from(e);
 
         log.error(response.toString());
 
         return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(BusinessException.class)
+    public ResponseEntity<?> handleMemberIdNotFoundException(BusinessException e) {
+        BusinessExceptionResponse response = BusinessExceptionResponse.from(e);
+
+        log.error(response.toString());
+
+        return new ResponseEntity<>(response, e.getErrorCode().getHttpStatus());
     }
 
 }
