+++ airdnb/src/main/java/team10/airdnb/exception/response/ValidExceptionResponse.java
@@ -1,15 +1,9 @@
 package team10.airdnb.exception.response;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 
-@Getter
-@RequiredArgsConstructor
-public class ValidExceptionResponse {
-    private final String errorField;
-    private final String errorMessage;
+public record ValidExceptionResponse(String errorField, String errorMessage) {
 
     public static ValidExceptionResponse from(MethodArgumentNotValidException e) {
         FieldError fieldError = e.getFieldError();
@@ -22,10 +16,4 @@ public class ValidExceptionResponse {
 
         return new ValidExceptionResponse(e.getClass().getSimpleName(), e.getMessage());
     }
-
-    @Override
-    public String toString() {
-        return String.format("%s : %s", errorField, errorMessage);
-    }
-
 }
