+++ airdnb/src/main/java/team10/airdnb/member/exception/MemberIdNotFoundException.java
@@ -1,0 +1,10 @@
+package team10.airdnb.member.exception;
+
+import team10.airdnb.exception.BusinessException;
+import team10.airdnb.error.ErrorCode;
+
+public class MemberIdNotFoundException extends BusinessException {
+    public MemberIdNotFoundException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
