--- airdnb/src/main/java/team10/airdnb/oauth/error/exception/EntityNotFoundException.java
@@ -1,12 +1,0 @@
-package team10.airdnb.oauth.error.exception;
-
-import team10.airdnb.oauth.error.ErrorCode;
-
-public class EntityNotFoundException extends BusinessException {
-
-    public EntityNotFoundException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-
-}
-
