+++ airdnb/src/main/java/team10/airdnb/oauth/exception/AuthenticationException.java
@@ -1,7 +1,8 @@
-package team10.airdnb.oauth.error.exception;
+package team10.airdnb.oauth.exception;
 
 
-import team10.airdnb.oauth.error.ErrorCode;
+import team10.airdnb.exception.BusinessException;
+import team10.airdnb.error.ErrorCode;
 
 public class AuthenticationException extends BusinessException {
