+++ airdnb/src/main/java/team10/airdnb/oauth/login/service/OauthLoginService.java
@@ -7,8 +7,8 @@ import org.springframework.transaction.annotation.Transactional;
 import team10.airdnb.member.constant.MemberType;
 import team10.airdnb.member.entity.Member;
 import team10.airdnb.member.service.MemberService;
-import team10.airdnb.oauth.jwt.dto.JwtTokenDto;
-import team10.airdnb.oauth.jwt.service.TokenManager;
+import team10.airdnb.jwt.dto.JwtTokenDto;
+import team10.airdnb.jwt.service.TokenManager;
 import team10.airdnb.oauth.login.dto.OauthLoginDto;
 import team10.airdnb.oauth.model.OAuthAttributes;
 import team10.airdnb.oauth.service.SocialLoginApiService;
