+++ airdnb/src/main/java/team10/airdnb/oauth/login/validator/OauthValidator.java
@@ -2,8 +2,8 @@ package team10.airdnb.oauth.login.validator;
 
 import org.springframework.stereotype.Service;
 import team10.airdnb.member.constant.MemberType;
-import team10.airdnb.oauth.error.ErrorCode;
-import team10.airdnb.oauth.error.exception.BusinessException;
+import team10.airdnb.error.ErrorCode;
+import team10.airdnb.exception.BusinessException;
 
 @Service
 public class OauthValidator {
